Extract page publication fields handler into ESM
This is not sitemap related and should be it's own module
