+++ app/views/alchemy/admin/pages/_sitemap.html.erb
@@ -31,6 +31,6 @@
      }
      <% end %>
    });
-   Alchemy.Sitemap.watchPagePublicationState();
+   Alchemy.PagePublicationFields();
  });
 </script>
