+++ app/views/alchemy/admin/pages/edit.html.erb
@@ -154,7 +154,7 @@
       if (!not_dirty) Alchemy.pleaseWaitOverlay(false);
       return not_dirty;
     });
-    Alchemy.Sitemap.watchPagePublicationState();
+    Alchemy.PagePublicationFields();
     Alchemy.PageLeaveObserver();
     Alchemy.ElementsWindow.init('<%= alchemy.admin_elements_path(page_version_id: @page_version.id) %>', {
       texts: {
